Updated admin for union, department and user
